+++ http4k-security-oauth/src/main/kotlin/org/http4k/security/oauth/server/ClientValidationFilter.kt
@@ -10,24 +10,25 @@ import org.http4k.core.HttpHandler
 import org.http4k.security.ResponseType
 
 class ClientValidationFilter(private val authoriseRequestValidator: AuthoriseRequestValidator,
-                             private val errorRenderer: ErrorRenderer,
+                             private val errorRenderer: AuthoriseRequestErrorRender,
                              private val extractor: AuthRequestExtractor) : Filter {
 
     override fun invoke(next: HttpHandler): HttpHandler =
-            {
-                if (!validResponseTypes.contains(it.query("response_type"))) {
-                    errorRenderer.response(UnsupportedResponseType(it.query("response_type").orEmpty()))
-                } else {
-                    extractor.extract(it).map { authorizationRequest ->
-                        when(val result = authoriseRequestValidator.validate(it, authorizationRequest)) {
-                            is Success -> next(result.value)
-                            is Failure -> errorRenderer.response(result.reason)
-                        }
-                    }.mapFailure(errorRenderer::response).get()
-                }
+        {
+            if (!validResponseTypes.contains(it.query("response_type"))) {
+                errorRenderer.errorFor(it, UnsupportedResponseType(it.query("response_type").orEmpty()))
+            } else {
+                extractor.extract(it).map { authorizationRequest ->
+                    when (val result = authoriseRequestValidator.validate(it, authorizationRequest)) {
+                        is Success -> next(result.value)
+                        is Failure -> errorRenderer.errorFor(it, result.reason)
+                    }
+                }.mapFailure { error -> errorRenderer.errorFor(it, error) }.get()
             }
+        }
 
     companion object {
         val validResponseTypes = ResponseType.values().map { it.queryParameterValue }
     }
 }
+