+++ CHANGELOG.md
@@ -5,6 +5,8 @@ changes with their rationale when appropriate:
 
 ### v3.218.0 (uncut)
 - [http4k-incubator] Next iteration of Servirtium JUnit extensions. Correct indexing of interactions.
+- [http4k-security-oauth] Authorisation rendering will now taking into account 'response_mode' of either query or fragment in responses and no longer just use the default fo the 'response_type'. H/T @tom
+- [http4k-security-oauth] [Breaking] Error responses in the authorise endpoint will actually redirect back to 'redirect_uri' assuming the validator correctly validates both the 'client_id' and 'redirect_uri' to be valid. H/T @tom
 
 ### v3.217.0
 - [all] Upgrade some dependency versions.